config/volumes/workload/pyproject.toml
[project]
    name = "workload"
    version = "0.1.0"
    description = "Add your description here"
    authors = [
        { name = "Michael Legleux", email = "legleux@gmail.com" }
    ]
    dependencies = [
    "xrpl-py",
    "requests",
    "matplotlib>=3.9.2",
    "pyqt6>=6.7.1",
]
    readme = "README.md"
    requires-python = ">= 3.11"
    license = { text = "MIT" }

[project.scripts]
    "workload" = "workload.app:main"

[build-system]
    requires = ["hatchling"]
    build-backend = "hatchling.build"

[tool.rye]
    managed = true
    dev-dependencies = [
        "ruff"
    ]

[tool.hatch.metadata]
    allow-direct-references = true

[tool.hatch.build.targets.wheel]
    packages = [
        "src/workload",
    ]
